requirements.txt
fastapi
uvicorn
transformers
torch
python-multipart
discord.py-self
tf-keras